Correcciones varias del modelo.

+++ src/main/java/com/quantumdev/integraservicios/model/Building.java
@@ -28,7 +28,7 @@ public class Building {
         name = "code_building",
         nullable = false
     )
-    private Byte code;
+    private Short code;
 
     @NotEmpty
     @ManyToOne(
